+++ ktan-processor/src/main/java/com/ktan/processor/KtanBuilder.kt
@@ -1,0 +1,6 @@
+package com.ktan.processor
+
+interface KtanBuilder<T> {
+
+    fun build(): T
+}