+++ ktan-processor/src/main/java/com/ktan/processor/extensions/KSValueParameterExt.kt
@@ -1,0 +1,16 @@
+package com.ktan.processor.extensions
+
+import com.google.devtools.ksp.symbol.KSValueParameter
+
+fun KSValueParameter.isNotKotlinPrimitive(): Boolean {
+
+    return when (type.element?.toString()) {
+        "String", "Int", "Short", "Number", "Boolean", "Byte", "Char", "Float", "Double", "Long", "Unit", "Any" -> false
+        else -> true
+    }
+}
+
+fun KSValueParameter.getPrimitiveTypeName(): String {
+
+    return type.element?.toString() ?: throw IllegalAccessException()
+}