+++ ktan-processor/src/main/java/com/ktan/processor/properties/KtanPropertyAdapter.kt
@@ -1,0 +1,16 @@
+package com.ktan.processor.properties
+
+import com.google.devtools.ksp.symbol.KSAnnotation
+
+interface KtanPropertyAdapter {
+
+    fun isMatch(annotations: Sequence<KSAnnotation>): Boolean
+
+    fun onInit(name: String)
+
+    fun onMutableIsPresent()
+
+    fun onRequiredIsPresent()
+
+    fun build()
+}